feat(RandBorrowed): Provide borrowed interface for rand compatibility
